clustering

+++ sbbi_clustering/clustering.py
@@ -2,13 +2,16 @@ import csv
 import glob
 import os
 import json
-import shutil
+import argparse
 
 
 #def get_recipe_nutrition():
 def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument("-c", "--country", help="country information", required=True)
+    args = parser.parse_args().country
     csv_columns = ['Total Fat', 'Saturated Fat', 'Cholesterol', 'Sodium', 'Potassium','Total Carbohydrates', 'Dietary Fiber', 'Protein', 'Sugars', 'Vitamin A', 'Vitamin C', 'Calcium', 'Iron', 'Thiamin', 'Niacin', 'Vitamin B6', 'Magnesium', 'Folate']
-    folders = glob.glob('Americe/*')
+    folders = glob.glob(args + '/*')
     for folder in folders:
         files = glob.glob(folder + '/*')
         for file in files:
@@ -30,14 +33,23 @@ def main():
                     dict_data.append(recipe['nutrition'])
 
             csv_file = "recipes_nutrition.csv"
-            try:
-                with open(csv_file, 'w') as csvfile:
-                    writer = csv.DictWriter(csvfile, fieldnames=csv_columns)
-                    writer.writeheader()
-                    for data in dict_data:
-                        writer.writerow(data)
-            except IOError:
-                print("I/O error")
+            if os.path.isfile(csv_file):
+                try:
+                    with open(csv_file, 'a+') as csvfile:
+                        writer = csv.DictWriter(csvfile, fieldnames=csv_columns)
+                        for data in dict_data:
+                            writer.writerow(data)
+                except IOError:
+                    print("I/O error")
+            else:
+                try:
+                    with open(csv_file, 'w') as csvfile:
+                        writer = csv.DictWriter(csvfile, fieldnames=csv_columns)
+                        writer.writeheader()
+                        for data in dict_data:
+                            writer.writerow(data)
+                except IOError:
+                    print("I/O error")
